Spotify-ytmusic/requirements.txt
flask
spotipy
ytmusicapi
python-dotenv
